refactor: set up initial typescript support
