+++ .gitignore
@@ -1,3 +1,4 @@
+build/
 
 # Created by https://www.toptal.com/developers/gitignore/api/node,jetbrains+all,visualstudiocode
 # Edit at https://www.toptal.com/developers/gitignore?templates=node,jetbrains+all,visualstudiocode
