+++ src/cmd/create-commands.js
@@ -25,7 +25,7 @@ const commands = [];
 const commandFiles = fs.readdirSync('./commands').filter((file) => file.endsWith('.js'));
 
 for (const file of commandFiles) {
-  const command = require(`./commands/${file}`);
+  const command = require(`${__dirname}/../commands/${file}`);
   commands.push(command.data.toJSON());
 }
