+++ src/cmd/delete-commands.js
