+++ src/commands/price.js
